+++ lib/routes.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'views/testing/3d_model_viewer_screen.dart';
-import 'package:learn_with_ar/views/dashboard_screen.dart';
-import 'package:learn_with_ar/views/login_screen.dart';
+import 'package:learn_with_ar/views/DashboardScreen.dart';
+import 'package:learn_with_ar/views/LoginScreen.dart';
 import 'views/testing/obj_modal_view.dart';
 import 'views/testing/render_object.dart';
-import 'package:learn_with_ar/views/uploadScreen.dart';
-import 'package:learn_with_ar/views/profileScreen.dart';
+import 'package:learn_with_ar/views/UploadScreen.dart';
+import 'package:learn_with_ar/views/ProfileScreen.dart';
 
 Map<String, Widget Function(BuildContext)> route_generator = {
   'login': (context) => LoginScreen(),
