+++ lib/views/LoginScreen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_login/flutter_login.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-//import 'dashboard_screen.dart';
+//import 'DashboardScreen.dart';
 
 class LoginScreen extends StatelessWidget {
   final _auth = FirebaseAuth.instance;
