+++ lib/views/UploadScreen.dart
@@ -1,6 +1,6 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/src/widgets/container.dart';
+import 'package:learn_with_ar/views/widgets/nav_bar.dart';
 import 'package:learn_with_ar/views/widgets/dashboard_grid_container.dart';
 
 class uploadScreen extends StatefulWidget {
